separate out pydantic config classes
