+++ src/dolphin/workflows/__init__.py
@@ -1,5 +1,4 @@
 """Tools for creating workflows and running displacement workflows."""
 
-from ._enums import *  # noqa
 from ._utils import *  # noqa
 from .config import *  # noqa
