+++ src/dolphin/workflows/config/__init__.py
@@ -1,0 +1,5 @@
+"""Configuration and validation models for (sub)workflows."""
+
+from ._common import *  # noqa
+from ._displacement import *  # noqa
+from ._enums import *  # noqa
